fix: type sink capture coordinates and skip inactive score tiles

+++ crates/memra-engine/src/dsv4_gpu.rs
@@ -15459,7 +15459,7 @@ impl Dsv4Gpu {
             unsafe {
                 observer(
                     st.dev,
-                    layer.il,
+                    layer.il as usize,
                     layer.ratio,
                     slots,
                     dpf!(vws.q, &stream),
